add folder edit function

+++ src/Convience.Web/Managent/src/app/modules/content-manage/file-manage/file-manage.component.ts
@@ -4,6 +4,7 @@ import { FileInfo } from '../model/fileInfo';
 import { NzModalService, NzModalRef, UploadFile, NzMessageService } from 'ng-zorro-antd';
 import { FileService } from 'src/app/services/file.service';
 import { FolderService } from 'src/app/services/folder.service';
+import { FormGroup, FormBuilder, Validators, FormControl } from '@angular/forms';
 
 @Component({
   selector: 'app-file-manage',
@@ -11,6 +12,8 @@ import { FolderService } from 'src/app/services/folder.service';
   styleUrls: ['./file-manage.component.scss']
 })
 export class FileManageComponent implements OnInit {
+
+  folderForm: FormGroup = new FormGroup({});
 
   fileInfoList: FileInfo[] = [];
 
@@ -25,27 +28,41 @@ export class FileManageComponent implements OnInit {
   @ViewChild('uploadFooterTpl', { static: true })
   uploadFooterTpl;
 
+  @ViewChild('folderContentTpl', { static: true })
+  folderContentTpl;
+
   modal: NzModalRef;
 
   fileList: UploadFile[] = [];
 
   uploading: boolean = false;
 
-  currentDirectory: string = '';
+  currentDirectory: string = '/';
+
+  emptyValidator = (control: FormControl): { [key: string]: any } | null => {
+    const reg = /^\s+/g;
+    let value = control.value?.replace(reg, '');
+    return value ? null : { 'notEmpty': true };
+  };
 
   constructor(private modalService: NzModalService,
     private messageService: NzMessageService,
     private fileService: FileService,
-    private folderService: FolderService) {
+    private folderService: FolderService,
+    private formBuilder: FormBuilder) {
   }
 
   ngOnInit(): void {
     this.refresh();
   }
 
-  dbc(dir) {
-    this.currentDirectory = dir;
-    this.folderList = dir.split('/').filter(e => e);
+  dbc(file: FileInfo) {
+    if (file.isDirectory) {
+      let dir = file.directory + '/' + file.fileName
+      this.folderList = dir.split('/').filter(e => e);
+      this.currentDirectory = '/' + this.folderList.join('/');
+      this.refresh();
+    }
   }
 
   refresh() {
@@ -82,9 +99,22 @@ export class FileManageComponent implements OnInit {
   }
 
   delete(fileInfo: FileInfo) {
-    this.fileService.delete(fileInfo.fileName, fileInfo.directory).subscribe(result => {
-      this.messageService.success("删除成功！");
-      this.refresh();
+    this.modalService.confirm({
+      nzTitle: '是否删除?',
+      nzContent: null,
+      nzOnOk: () => {
+        if (fileInfo.isDirectory) {
+          this.folderService.deleteFolder(fileInfo.directory, fileInfo.fileName).subscribe(result => {
+            this.messageService.success("删除成功！");
+            this.refresh();
+          });
+        } else {
+          this.fileService.delete(fileInfo.fileName, fileInfo.directory).subscribe(result => {
+            this.messageService.success("删除成功！");
+            this.refresh();
+          });
+        }
+      }
     });
   }
 
@@ -96,5 +126,46 @@ export class FileManageComponent implements OnInit {
       a.download = fileInfo.fileName;
       a.click();
     });
+  }
+
+  createFolder() {
+    this.modal = this.modalService.create({
+      nzTitle: '创建文件夹',
+      nzContent: this.folderContentTpl,
+      nzFooter: null
+    });
+    this.folderForm = this.formBuilder.group({
+      folderName: [null, [Validators.required, Validators.maxLength(15),
+      Validators.pattern('^[^\\\\/:\*\?""<>|]{1,120}$'), this.emptyValidator]]
+    });
+  }
+
+  submit() {
+    for (const i in this.folderForm.controls) {
+      this.folderForm.controls[i].markAsDirty();
+      this.folderForm.controls[i].updateValueAndValidity();
+    }
+    if (this.folderForm.valid) {
+      this.folderService.addFolder(this.currentDirectory, this.folderForm.value['folderName']).subscribe(result => {
+        this.messageService.success("文件夹创建成功！");
+        this.refresh();
+        this.modal.close();
+      });
+    }
+  }
+
+  cancel() {
+    this.modal.close();
+  }
+
+  navgiateFolder(index) {
+    if (index == -1) {
+      this.currentDirectory = '/';
+      this.folderList = [];
+    } else {
+      this.folderList = this.folderList.slice(0, index + 1);
+      this.currentDirectory = '/' + this.folderList.join('/');
+    }
+    this.refresh();
   }
 }
